Fix querying summary variants
