+++ dae/dae/backends/schema2/family_builder.py
@@ -5,6 +5,8 @@ logger = logging.getLogger(__name__)
 
 
 class FamilyQueryBuilder(BaseQueryBuilder):
+    """Build queries related to family variants"""
+
     def __init__(
         self,
         dialect: Dialect,
@@ -21,7 +23,7 @@ class FamilyQueryBuilder(BaseQueryBuilder):
         gene_models=None,
         do_join_affected=False,
     ):
-
+        #pylint: disable=too-many-arguments
         self.family_variant_table = family_variant_table
         self.summary_allele_table = summary_allele_table
 
@@ -42,31 +44,26 @@ class FamilyQueryBuilder(BaseQueryBuilder):
         self.do_join_affected = do_join_affected
         self.families = families
 
-    def _where_accessors(self):
-        return super()._where_accessors()
-
     def _query_columns(self):
         self.select_accessors = {
-            "bucket_index": f"sa.bucket_index",
-            "summary_index": f"sa.summary_index",
-            "family_index": f"fa.family_index",
-            "family_id": f"fa.family_id",
+            "bucket_index": "sa.bucket_index",
+            "summary_index": "sa.summary_index",
+            "family_index": "fa.family_index",
+            "family_id": "fa.family_id",
             "summary_data": "sa.summary_data",
-            "family_data": f"fa.family_data",
+            "family_data": "fa.family_data",
         }
 
         columns = list(self.select_accessors.values())
         return columns
 
-    def build_from(self):
-        return super().build_from()
-
-    def build_join(self, genes=None, effect_types=None):
-
+    def _build_join(self, genes=None, effect_types=None):
         join_clause = ""
 
         if self.do_join_affected:
-            join_clause = f"JOIN {self.dialect.build_table_name(self.pedigree_table, self.db)} as pedigree\n"
+            pedigree_table = self.dialect.build_table_name(self.pedigree_table,
+                                                           self.db)
+            join_clause = f"JOIN {pedigree_table} as pedigree\n"
 
         if genes is not None or effect_types is not None:
             effect_gene_abs = self.where_accessors["effect_gene"]
@@ -79,13 +76,25 @@ class FamilyQueryBuilder(BaseQueryBuilder):
 
         self._add_to_product(join_clause)
 
-    def build_group_by(self):
+    def _build_from(self):
+        # implicit join on family_allele and summary variants table
+        from_clause = f"""\n FROM
+        {self.dialect.build_table_name(self.summary_allele_table, self.db)} AS sa
+        JOIN 
+        {self.dialect.build_table_name(self.family_variant_table, self.db)} AS fa
+        ON (fa.summary_index = sa.summary_index AND
+        fa.bucket_index = sa.bucket_index AND
+        fa.allele_index = sa.allele_index)""".rstrip()
+
+        self._add_to_product(from_clause)
+
+    def _build_group_by(self):
         pass
 
-    def build_having(self, **kwargs):
+    def _build_having(self, **kwargs):
         pass
 
-    def build_where(
+    def _build_where(
         self,
         regions=None,
         genes=None,
@@ -103,6 +112,7 @@ class FamilyQueryBuilder(BaseQueryBuilder):
         return_unknown=None,
         **kwargs,
     ):
+        #pylint: disable=too-many-arguments,too-many-locals
         if self.summary_allele_table:
             inheritance = None
         where_clause = self._base_build_where(
